+++ sleek_storage/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 * BREAKING: `Box.watch` now emits `null` when value is deleted or when box is cleared.
 * BREAKING: `fromJson` now also provides the key.
+* New: `SleekBox` is now a lazy `Iterable`.
 * New `Box.containsKey` method.
 * New optional `defaultValue` parameter on `SleekStorage.value()` method.
 * Fix `Box.putAll` method.